Refactored ProjectController
