+++ packages/api/AlvTimeWebApi/ServiceRegistrator.cs
@@ -41,7 +41,7 @@ namespace AlvTimeWebApi
             services.AddScoped<IHourRateStorage, HourRateStorage>();
             services.AddScoped<HourRateService>();
             services.AddScoped<IProjectStorage, ProjectStorage>();
-            services.AddScoped<ProjectCreator>();
+            services.AddScoped<ProjectService>();
             services.AddScoped<ICustomerStorage, CustomerStorage>();
             services.AddScoped<CustomerService>();
             services.AddScoped<IEconomyStorage, EconomyStorage>();
